requirements.txt
Pillow==10.2.0
protobuf==4.25.2
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.30.0
google.generativeai